requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.10
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.3.0
django-filter>=23.3
python-dotenv>=1.0.0
